requirements.txt
fastapi
uvicorn
openai-whisper
yt-dlp
requests 